backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.5",
    "google-genai>=1.62.0",
    "openai>=2.17.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.32",
    "uvicorn[standard]>=0.40.0",
]
